% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{call_merge}
\alias{call_merge}
\title{Perform full join between two data tables}
\usage{
call_merge(dt1, dt2)
}
\arguments{
\item{dt1:}{The first data table}

\item{dt2:}{The second data table}
}
\value{
A data.table
}
\description{
This function takes input as two data tables and perform full join between the two.
The join keys are all the common columns between the two data table
}

