% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{column_inlinebox}
\alias{column_inlinebox}
\title{Creates column of inline box plot}
\usage{
column_inlinebox(edadata)
}
\arguments{
\item{data:}{The data table or data frame}
}
\value{
A frame
}
\description{
This function takes input as a data table or data frame.
It creates an inline Box plot using Sparkline
}
