% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{complete_summary}
\alias{complete_summary}
\title{Creates the complete summary of data}
\usage{
complete_summary(data)
}
\arguments{
\item{data:}{The data table or data frame}
}
\value{
A frame
}
\description{
This function takes input as a data table or data frame.
The function creates complete summary of the data.
}
