% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_inlinebox}
\alias{create_inlinebox}
\title{Creates inline box plot}
\usage{
create_inlinebox(x)
}
\arguments{
\item{data:}{The data table or data frame}
}
\value{
A frame
}
\description{
This function takes input as a data table or data frame.
It creates an inline Box plot using Sparkline
}
