% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{date_summary}
\alias{date_summary}
\title{Creates summary for date variable}
\usage{
date_summary(data)
}
\arguments{
\item{data:}{The data table or data frame}
}
\value{
A data frame
}
\description{
This function takes input as a data table or data frame.
The function creates summary for the date variable.
}

