% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{exploratory_report}
\alias{exploratory_report}
\title{Returns a report.}
\usage{
exploratory_report(data, filename = NULL,
  output_format = "html_document")
}
\arguments{
\item{data:}{The data frame}

\item{filename:}{String name for the report name.}

\item{output_format:}{Output Format.}
}
\value{
A Report for the Data Analysis
}
\description{
This function takes input data.
The function calculates the number of missing values in the vector.
}
