% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{factor_summary}
\alias{factor_summary}
\title{Creates summary for factor variable}
\usage{
factor_summary(data)
}
\arguments{
\item{data:}{The data table or data frame}
}
\value{
A data frame
}
\description{
This function takes input as a data table or data frame.
The function creates summary for the factor variable.
}
