% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{frequent_count}
\alias{frequent_count}
\title{Returns a data frame with requenct count of date, factor, and character variables.}
\usage{
frequent_count(data, var)
}
\arguments{
\item{x:}{The data frame}

\item{variable:}{String name of the variable.}
}
\value{
A data table
}
\description{
This function takes input data.
The function calculates the number of missing values in the vector.
}
