% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{group_bin}
\alias{group_bin}
\title{Group Column based on bin size}
\usage{
group_bin(data, x, y = NULL, breaks = NULL, ...)
}
\arguments{
\item{data}{Input data table}

\item{x:}{A single numeric column to get top n and min n numbers}

\item{y:}{Column to summarize default is input column}

\item{break:}{bin size of data}
}
\value{
returns a data.frame with two columns with min and max numbers
}
\description{
This function groups columns based on bin size of numeric data. 
Ideally The breaks should start from -Inf and end at +Inf to avoid Nan's. 
Otherwise we can see see Nan's. Default bin's are calculated based on quartiles of the data
}
