% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{is_date}
\alias{is_date}
\title{Check Date Type}
\usage{
is_date(x)
}
\arguments{
\item{x:}{A single column of data table}
}
\value{
Logical True if the input column is of date type else False
}
\description{
This function checkes if the input column is of date type.
}
