% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{join_data}
\alias{join_data}
\title{Compares two data tables}
\usage{
join_data(dt1, dt2)
}
\arguments{
\item{dt1:}{The first data table}

\item{dt2:}{The second data table}
}
\value{
A data.table
}
\description{
This function takes input as two data tables.
The function checks if any one data.table is empty and returns the non-empty data.table
If both the data tables are non-empty then it call the call_merge to perform full join on the two data tables.
}
