% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{log_summary}
\alias{log_summary}
\title{Creates summary for logical variable}
\usage{
log_summary(data)
}
\arguments{
\item{data:}{The data table or data frame}
}
\value{
A data frame
}
\description{
This function takes input as a data table or data frame.
The function creates summary for the logical variable.
}

