% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{maxmin_count}
\alias{maxmin_count}
\title{Returns a data frame with maximium 5 and minimum 5.}
\usage{
maxmin_count(data, var)
}
\arguments{
\item{x:}{The data frame}

\item{variable:}{String name of the variable.}
}
\value{
A data frame
}
\description{
This function takes input data.
The function calculates the number of missing values in the vector.
}
