% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{miss_sum}
\alias{miss_sum}
\title{Returns the number or percentage of missing value}
\usage{
miss_sum(x, count)
}
\arguments{
\item{x:}{The vector}

\item{count:}{Scalar quantity either 1 or number of rows in the data.}
}
\value{
A vector
}
\description{
This function takes input vector.
The function calculates the number of missing values in the vector.
}

