% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{num_summary}
\alias{num_summary}
\title{Creates summary for numeric variable}
\usage{
num_summary(data)
}
\arguments{
\item{data:}{The data table or data frame}
}
\value{
A data frame
}
\description{
This function takes input as a data table or data frame.
The function creates summary for the numeric variable.
}

