% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Returns a scatter plot object.}
\usage{
scatter_plot(data, x_var, y_var)
}
\arguments{
\item{data:}{The data frame}

\item{x_var:}{String name of the variable.}

\item{y_var:}{String name of the variable.}
}
\value{
A rbokeh plot object
}
\description{
This function takes input data.
The function calculates the number of missing values in the vector.
}

