% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{start_shiny}
\alias{start_shiny}
\title{The functions executes the dashboard using shiny to explore the data set.}
\usage{
start_shiny(port = getOption("shiny.port"))
}
\arguments{
\item{port}{The TCP port that the application should listen on. If the
\code{port} is not specified, and the \code{shiny.port} option is set (with
\code{options(shiny.port = XX)}), then that port will be used. Otherwise,
use a random port.}
}
\description{
The functions executes the dashboard using shiny to explore the data set.
}
