% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{top_min_n}
\alias{top_min_n}
\title{Check Date Type}
\usage{
top_min_n(data, x_var, n = 5)
}
\arguments{
\item{x:}{A single numeric column to get top n and min n numbers}

\item{n:}{count of top numbers to return}
}
\value{
returns a data.frame with two columns with min and max numbers
}
\description{
This function checkes if the input column is of date type.
}
