% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{unique_sum}
\alias{unique_sum}
\title{Returns number of unique characters}
\usage{
unique_sum(x)
}
\arguments{
\item{x:}{A character, date, or factor vector}
}
\value{
A vector
}
\description{
This function takes input as vector or column of data table or data frame.
The function returns the number of unique elements in the vector.
}
