% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{zeros_sum}
\alias{zeros_sum}
\title{Returns the number or percentage of zeros value}
\usage{
zeros_sum(x, count)
}
\arguments{
\item{x:}{The vector}

\item{count:}{Scalar quantity either 1 or number of rows in the data.}
}
\value{
A vector
}
\description{
This function takes input vector.
The function calculates the number of zeros values in the vector.
}

