% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{top_min_n}
\alias{top_min_n}
\title{Find Max and Min of the Numeric Columns}
\usage{
top_min_n(data, x_var, n = 5)
}
\arguments{
\item{data:}{Input data table}

\item{x:}{A single numeric column to get top n and min n numbers}

\item{n:}{count of top numbers to return}
}
\value{
returns a data.frame with two columns with min and max numbers
}
\description{
This function finds max n and min n data of the numeric columns
}
